refactor(#4): move customer open api related things to specif dirs
