+++ src/Shared/Application/OpenApi/Builder/ResponseBuilderInterface.php
@@ -1,8 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Shared\Application\OpenApi\Builder;
 
 interface ResponseBuilderInterface
 {
-
 }
