+++ src/Shared/Application/OpenApi/Factory/Request/Customer/UpdateCustomerRequestFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Shared\Application\OpenApi\Factory\Request;
+namespace App\Shared\Application\OpenApi\Factory\Request\Customer;
 
 use App\Shared\Application\OpenApi\Builder\Parameter;
 use App\Shared\Application\OpenApi\Builder\RequestBuilder;
