+++ src/Shared/Application/OpenApi/Factory/Response/Customer/CustomerDeletedResponseFactory.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\Shared\Application\OpenApi\Factory\Response;
+namespace App\Shared\Application\OpenApi\Factory\Response\Customer;
 
 use ApiPlatform\OpenApi\Model\Response;
 use App\Shared\Application\OpenApi\Builder\ResponseBuilder;
+use App\Shared\Application\OpenApi\Factory\Response\ResponseFactoryInterface;
 
 final class CustomerDeletedResponseFactory implements ResponseFactoryInterface
 {
