+++ src/Shared/Application/OpenApi/Factory/Response/Customer/CustomerResponseFactory.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace App\Shared\Application\OpenApi\Factory\Response;
+namespace App\Shared\Application\OpenApi\Factory\Response\Customer;
 
 use ApiPlatform\OpenApi\Model\Response;
 use App\Shared\Application\OpenApi\Builder\Parameter;
 use App\Shared\Application\OpenApi\Builder\ResponseBuilderInterface;
+use App\Shared\Application\OpenApi\Factory\Response\ResponseFactoryInterface;
 
 abstract class CustomerResponseFactory implements ResponseFactoryInterface
 {
